Moving set_default_dataset_id to _implicit_environ module.
Also splitting out behavior into a get_default_dataset_id method
and then the set_default_dataset_id around in.
