+++ gcloud/datastore/_implicit_environ.py
@@ -12,12 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Module to provide implicit behavior based on enviroment.
+"""Module to provide implicit behavior based on environment.
 
 Acts as a mutable namespace to allow the datastore package to
-imply the current dataset ID and connection from the enviroment.
+imply the current dataset ID and connection from the environment.
 """
 
+import os
 import socket
 
 from six.moves.http_client import HTTPConnection  # pylint: disable=F0401
@@ -28,14 +29,17 @@ except ImportError:
     app_identity = None
 
 
-class Enviroment(object):
+_DATASET_ENV_VAR_NAME = 'GCLOUD_DATASET_ID'
+
+
+class Environment(object):
     """Container for environment settings."""
 
     dataset_id = None
-    """Attribute to allow persistent implied dataset ID from enviroment."""
+    """Attribute to allow persistent implied dataset ID from environment."""
 
     connection = None
-    """Attribute to allow persistent implied connection from enviroment."""
+    """Attribute to allow persistent implied connection from environment."""
 
 
 def app_engine_id():
@@ -84,4 +88,53 @@ def compute_engine_id():
         connection.close()
 
 
-DEFAULT_ENVIRON = Enviroment()
+def get_default_dataset_id(dataset_id=None):
+    """Get default dataset ID either explicitly or implicitly as fall-back.
+
+    In implicit case, supports three cases. In order of precedence, the
+    implicit cases are:
+    - GCLOUD_DATASET_ID environment variable
+    - Google App Engine application ID
+    - Google Compute Engine project ID (from metadata server)
+
+    :type dataset_id: string
+    :param dataset_id: Optional. The dataset ID to use as default.
+
+    :rtype: string or ``NoneType``
+    :returns: The inferred dataset or None.
+    """
+    if dataset_id is None:
+        dataset_id = os.getenv(_DATASET_ENV_VAR_NAME)
+
+    if dataset_id is None:
+        dataset_id = app_engine_id()
+
+    if dataset_id is None:
+        dataset_id = compute_engine_id()
+
+    return dataset_id
+
+
+def set_default_dataset_id(dataset_id=None):
+    """Set default dataset ID either explicitly or implicitly as fall-back.
+
+    In implicit case, supports three cases. In order of precedence, the
+    implicit cases are:
+    - GCLOUD_DATASET_ID environment variable
+    - Google App Engine application ID
+    - Google Compute Engine project ID (from metadata server)
+
+    :type dataset_id: string
+    :param dataset_id: Optional. The dataset ID to use as default.
+
+    :raises: :class:`EnvironmentError` if no dataset ID was implied.
+    """
+    dataset_id = get_default_dataset_id(dataset_id=dataset_id)
+
+    if dataset_id is not None:
+        DEFAULT_ENVIRON.dataset_id = dataset_id
+    else:
+        raise EnvironmentError('No dataset ID could be inferred.')
+
+
+DEFAULT_ENVIRON = Environment()
