+++ regression/populate_datastore.py
@@ -17,9 +17,10 @@
 from six.moves import zip
 
 from gcloud import datastore
+from gcloud.datastore import _implicit_environ
 
 
-datastore._DATASET_ENV_VAR_NAME = 'GCLOUD_TESTS_DATASET_ID'
+_implicit_environ._DATASET_ENV_VAR_NAME = 'GCLOUD_TESTS_DATASET_ID'
 datastore.set_defaults()
